requirements.txt
aiodebug==2.3.0
aiohttp==3.8.6
aioprocessing==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
fastapi==0.104.1
frozenlist==1.4.0
graphql-core==3.2.3
h11==0.14.0
hexbytes==0.3.1
idna==3.4
jsonpickle==3.0.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
mergedeep==1.3.4
multidict==6.0.4
parsimonious==0.9.0
prometheus-async==22.2.0
prometheus-client==0.18.0
protobuf==4.25.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.12.0
sgqlc==16.3
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0
web3==6.15.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
